[Doc] Add accurate documentation for `Differentiable` protocol.
Add accurate API documentation for the `Differentiable` protocol in
`stdlib/public/core/Differentiable-Documentation.swift`.

That file contains the ideal implementation of the `Differentiable` protocol
(in the absence of generic signature bugs) and exists only to be processed
during API doc generation.

The actual implementation of `Differentiable` is marked with `:nodoc:` so that
it is ignored during API doc generation.