[WIP] Toolchain size reduction attempt, using `copy_file_preserving_symlinks`.
Fails because `ln -s` peers through soft links for some reason.

```
$ ls build/buildbot_linux/swift-linux-x86_64/lib/swift/linux/libtensorflow.so*
lrwxrwxrwx   23 libtensorflow.so -> libtensorflow.so.1.14.0
lrwxrwxrwx   23 libtensorflow.so.1 -> libtensorflow.so.1.14.0
-rwxr-xr-x 199M libtensorflow.so.1.14.0

$ ls swift/swift-nightly-install/usr/lib/swift/linux/libtensorflow.so*
lrwxrwxrwx   23 libtensorflow.so -> libtensorflow.so.1.14.0
-rwxr-xr-x 199M libtensorflow.so.1.14.0
```

This causes a linker error when running `swift test` for tensorflow/swift-apis:
```
$ swift test
[86/86] Linking TensorFlowPackageTests.xctest
/home/danielzheng/swift-apis/.build/x86_64-unknown-linux/debug/TensorFlowPackageTests.xctest: error while loading shared libraries: libtensorflow.so.1: cannot open shared object file: No such file or directory
```

`objdump -x TensorFlowPackageTests.xctest` shows that `libtensorflow.so.1` is a dependency for some reason.
Perhaps that can be changed to `libtensorflow.so` - I'm not sure what controls that though.